+++ trview.shaders/trview.shaders.vcxproj.filters
@@ -1,0 +1,9 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<Project ToolsVersion="4.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
+  <ItemGroup>
+    <FxCompile Include="level_pixel_shader.hlsl" />
+    <FxCompile Include="level_vertex_shader.hlsl" />
+    <FxCompile Include="ui_vertex_shader.hlsl" />
+    <FxCompile Include="ui_pixel_shader.hlsl" />
+  </ItemGroup>
+</Project>
