+++ trview/trview.vcxproj.filters
@@ -160,6 +160,9 @@
     <ClInclude Include="LevelSwitcher.h">
       <Filter>Program</Filter>
     </ClInclude>
+    <ClInclude Include="ResourceHelper.h">
+      <Filter>Graphics</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="Viewer.cpp">
@@ -261,6 +264,9 @@
     <ClCompile Include="LevelSwitcher.cpp">
       <Filter>Program</Filter>
     </ClCompile>
+    <ClCompile Include="ResourceHelper.cpp">
+      <Filter>Graphics</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ResourceCompile Include="trview.rc">
@@ -309,19 +315,23 @@
     </Image>
   </ItemGroup>
   <ItemGroup>
-    <FxCompile Include="level_pixel_shader.hlsl">
-      <Filter>Graphics</Filter>
-    </FxCompile>
-    <FxCompile Include="level_vertex_shader.hlsl">
-      <Filter>Graphics</Filter>
-    </FxCompile>
-  </ItemGroup>
-  <ItemGroup>
     <Text Include="Controls.txt" />
   </ItemGroup>
   <ItemGroup>
     <None Include="texturelist.txt">
       <Filter>Win\Resource Files</Filter>
+    </None>
+    <None Include="resources\level_pixel_shader.cso">
+      <Filter>Resources</Filter>
+    </None>
+    <None Include="resources\level_vertex_shader.cso">
+      <Filter>Resources</Filter>
+    </None>
+    <None Include="resources\ui_pixel_shader.cso">
+      <Filter>Resources</Filter>
+    </None>
+    <None Include="resources\ui_vertex_shader.cso">
+      <Filter>Resources</Filter>
     </None>
   </ItemGroup>
 </Project>
